% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc.clin.R
\docType{data}
\name{acc.clin}
\alias{acc.clin}
\title{acc.clin}
\format{
92 rows and 9 columns
}
\usage{
acc.clin
}
\description{
Dataset containing clinical information of ACC samples. Samples are on rows and features are on columns.
}
\keyword{datasets}
