% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotated.genes.R
\docType{data}
\name{annotated.genes}
\alias{annotated.genes}
\title{DataFrame containing genes annotated in GO and KEGG databases}
\format{
A dataframe with 10760 rows and 2 columns.
}
\usage{
data(annotated.genes)
}
\description{
This dataframe contains all genes annotated in GO and KEGG databases and will be used
to filter non-annotated genes with a double goal:
1. Speed up the entire classification process.
2. Remove genes that are not annotated and will not contribute to an enrichment.
}
\keyword{datasets}
