% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{auroc.FC.consensus}
\alias{auroc.FC.consensus}
\title{Compute AUROC and Fold Change for Consensus Genes}
\usage{
auroc.FC.consensus(df, cons_genes, class)
}
\arguments{
\item{df}{A data frame containing gene expression data with rows as samples and columns as genes.}

\item{cons_genes}{A character vector of selected consensus genes to be analyzed.}

\item{class}{A vector of class labels (0 or 1) corresponding to the samples in `df`.}
}
\value{
A ggplot object displaying the AUROC values for each gene, with fold change color-coding
and error bars for the confidence interval of AUROC.
}
\description{
This function computes the Area Under the Receiver Operating Characteristic curve (AUROC)
and the Fold Change (FC) for each gene in a set of consensus genes. It returns a plot of
AUROC values for each gene, with error bars representing the confidence interval for each AUROC
and colors indicating the Fold Change values.
}
\details{
The function uses the \code{\link[pROC]{roc}} function to compute the AUROC for each gene in the
consensus set and calculates the Fold Change manually as the difference in the mean expression
between the two classes (class 1 and class 0).

The output plot orders genes based on their mean AUROC and color-codes the points according to
their corresponding Fold Change. The error bars represent the 95% confidence intervals for the
AUROC values.
}
\examples{
\dontrun{
# Example usage:
# Assuming 'expression_data' is a data frame with
# gene expression data and 'labels' is a vector of class labels
# cons_genes <- c("Gene1", "Gene2", "Gene3") # Example of consensus genes
# result_plot <- auroc.FC.consensus(expression_data, cons_genes, labels)
# print(result_plot)
}

}
