% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_testConsensus.R
\name{auroc_fc_calculation}
\alias{auroc_fc_calculation}
\title{AUROC and Fold Change Calculation}
\usage{
auroc_fc_calculation(df, cons_genes, class)
}
\arguments{
\item{df}{A data frame where rows represent samples and columns represent genes.}

\item{cons_genes}{A character vector containing the names of the consensus genes to be tested.}

\item{class}{A vector of class labels for each sample (binary, typically 0 and 1).}
}
\value{
A data frame with the following columns:
  - `gene`: The name of each consensus gene.
  - `auroc`: The AUROC value for each gene.
  - `auroc_upper`: The upper limit of the AUROC confidence interval.
  - `auroc_lower`: The lower limit of the AUROC confidence interval.
  - `FC`: The Fold Change for each gene, calculated as the difference in mean expression between the two classes.
}
\description{
Computes the Area Under the Receiver Operating Characteristic (AUROC) curve and Fold Change for each gene in a given set of consensus genes.
}
\details{
This function calculates the AUROC for each gene using the pROC package. The Fold Change is computed as the difference in mean expression between the two classes.
}
\examples{
\dontrun{
auroc_results <- auroc_fc_calculation(df = expression_data, cons_genes = gene_list, class = labels)
}

}
