% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{bar.plot}
\alias{bar.plot}
\title{Generate Bar Plot for Enrichment Results}
\usage{
bar.plot(
  enrichment_results,
  pcutoff = 0.05,
  low.col = "indianred",
  high.col = "lightblue"
)
}
\arguments{
\item{enrichment_results}{A data frame containing enrichment results with columns 'Description' and 'p.adjust'.}

\item{pcutoff}{Numeric specifying the adjusted p-value cutoff for highlighting significance (default: 0.05).}

\item{low.col}{Character specifying the color for low p-values (default: "indianred").}

\item{high.col}{Character specifying the color for high p-values (default: "lightblue").}
}
\value{
A bar plot visualizing enrichment results based on adjusted p-values.
}
\description{
This function generates a bar plot visualizing enrichment results based on adjusted p-values.
Each bar represents a different enriched module or pathway, ordered by significance.
}
\details{
The function orders the enrichment results by adjusted p-values and creates a bar plot where each bar represents
a module or pathway. Bars are colored based on the adjusted p-value, using a gradient from 'low.col' to 'high.col'.
A vertical dashed line is added at -log10(pcutoff) for visualizing the significance cutoff.
}
\examples{
\dontrun{
# Example usage of bar.plot function
# Assuming enrichment_results is a data frame with columns 'Description' and 'p.adjust'
bar.plot(enrichment_results = enrichment_results, pcutoff = 0.05)
bar.plot(enrichment_results = go_grouped_signif)
bar.plot(enrichment_results = go_grouped_signif[go_grouped_signif$p.adjust < 0.08, ])
bar.plot(enrichment_results = kegg_res)
}

}
