% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_preProcess.R
\name{batch.pca.plot}
\alias{batch.pca.plot}
\title{Generate PCA plots before and after batch correction}
\usage{
batch.pca.plot(data.before, data.after, batch, metadata)
}
\arguments{
\item{data.before}{A data frame containing the dataset before batch correction. The last column should not be included in the PCA.}

\item{data.after}{A data frame containing the dataset after batch correction. The last column should not be included in the PCA.}

\item{batch}{A character string specifying the column name in `metadata` that contains batch information.}

\item{metadata}{A data frame containing metadata for the samples, including batch information.}
}
\value{
A `ggplot2` object showing PCA plots before and after batch correction, with points colored by batch.
}
\description{
This function performs Principal Component Analysis (PCA) on datasets before and after batch correction
and visualizes the results with a scatter plot, coloring points by batch.
}
\details{
The function applies PCA separately to `data.before` and `data.after`, retaining the first two principal components.
The resulting PCA scores are plotted, with different colors representing batches. The plot includes density
distributions of PC1 and PC2 along the margins for additional visualization of batch effects.
}
\examples{
\dontrun{
# Example usage of batch.pca.plot
pp <- preProcess(df.count = acc.count, df.clin = acc.clin,
                batch = "batch", covar.mod = "covar")
batch.pca.plot(data.before =  pp@processed$normalized,
               data.after = pp@processed$adjusted.data,
               metadata = pp@metadata, batch = "batch")
}

}
