% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_preProcess.R
\name{batch_pvca_plot}
\alias{batch_pvca_plot}
\title{Generate PVCA plots before and after batch correction}
\usage{
batch_pvca_plot(data.before, data.after, metadata, class, batch, covar)
}
\arguments{
\item{data.before}{A data frame containing the dataset before batch correction.}

\item{data.after}{A data frame containing the dataset after batch correction.}

\item{metadata}{A data frame containing metadata for the samples, including batch and class information.}

\item{class}{A character string specifying the column name in `metadata` that contains class information (e.g., biological condition).}

\item{batch}{A character string specifying the column name in `metadata` that contains batch information.}

\item{covar}{A character string (or NULL) specifying an additional covariate to include in the PVCA model.}
}
\value{
A `ggplot2` object displaying the PVCA results, comparing variance components before and after batch correction.
}
\description{
This function performs Principal Variance Component Analysis (PVCA) to assess the proportion
of variance attributed to batch effects and other factors before and after batch correction.
}
\details{
The function applies PVCA separately to `data.before` and `data.after` to evaluate
the variance explained by batch effects, class, and additional covariates. The results
are visualized as a line plot showing the proportion of variance for each effect before and after correction.
}
\examples{
\dontrun{
# Example usage of batch_pvca_plot
pp <- preProcess(df.count = acc.count, df.clin = acc.clin,
                batch = "batch", covar.mod = "covar")
batch_pvca_plot(data.before =  pp@processed$normalized,
                data.after = pp@processed$adjusted.data,
                metadata = pp@metadata,
                batch = "batch", class = "class", covar = "covar")
}

}
