% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_runClassifiers.R
\name{check_required_pkgs}
\alias{check_required_pkgs}
\title{Check if required packages are installed}
\usage{
check_required_pkgs(pkgs)
}
\arguments{
\item{pkgs}{Character vector of package names to check.}
}
\value{
Invisibly TRUE if all packages installed, otherwise stops with a CLI alert.
}
\description{
Check if required packages are installed
}
\keyword{internal}
