% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{circos.plot}
\alias{circos.plot}
\title{Generate Circos Plot for Enrichment Results}
\usage{
circos.plot(
  enrichment_results,
  palette_genes = "Set2",
  palette_modules = "Paired",
  transparency = 0.5,
  facing = "clockwise",
  cex = 0.7,
  legend = TRUE,
  legend_title = "circos_plot_legend"
)
}
\arguments{
\item{enrichment_results}{A data frame containing enrichment results with columns 'ID' and 'Description'.}

\item{palette_genes}{Character specifying the color palette for genes (default: "Set2").}

\item{palette_modules}{Character specifying the color palette for modules or pathways (default: "Paired").}

\item{transparency}{Numeric specifying the transparency level for the plot elements (default: 0.5).}

\item{facing}{Character specifying the facing direction of text labels ("clockwise" or "counterclockwise", default: "clockwise").}

\item{cex}{Numeric specifying the character expansion factor for text labels (default: 0.7).}

\item{legend}{Logical indicating whether to include a legend in the plot (default: TRUE).}

\item{legend_title}{Character specifying the title for the legend (default: "circos_plot_legend").}
}
\value{
A Circos plot visualizing enrichment results based on genes and modules or pathways.
}
\description{
This function generates a Circos plot visualizing enrichment results using the chordDiagram function
from the circlize package. It plots genes and their enriched modules or pathways based on enrichment
results provided.
}
\details{
The function prepares data for the Circos plot by converting enrichment results into a matrix format
suitable for chordDiagram. It then creates a Circos plot using chordDiagram from circlize, with genes
and modules/pathways represented by different colors. If legend is enabled, it adds a legend to the plot
showing the modules or pathways and their descriptions.
}
\examples{
\dontrun{
# Example usage of circos.plot function
# Assuming enrichment_results is a data frame with columns 'ID' and 'Description'
circos.plot(
    enrichment_results = enrichment_results,
    palette_genes = "Set2", palette_modules = "Paired"
)
}

}
