% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrAnalysis.R
\name{enrich.GO}
\alias{enrich.GO}
\title{enrich.GO}
\usage{
enrich.GO(
  gene_list,
  keyType = "SYMBOL",
  pval = 0.05,
  ont = "BP",
  groupGO = TRUE,
  signif = TRUE,
  level = 4
)
}
\arguments{
\item{gene_list}{Character vector of gene symbols or IDs for which GO enrichment analysis is performed.}

\item{keyType}{Character specifying the type of gene identifier used in gene_list (default: "SYMBOL").}

\item{pval}{Numeric specifying the p-value cutoff for enriched GO terms (default: 0.05).}

\item{ont}{Character specifying the ontology ("BP" for Biological Process, "MF" for Molecular Function, "CC" for Cellular Component) (default: "BP").}

\item{groupGO}{Logical indicating whether to group similar GO terms (default: TRUE).}

\item{signif}{Logical indicating whether to filter significantly enriched GO terms based on p-value cutoff (default: TRUE).}

\item{level}{Integer specifying the depth level to group GO terms if groupGO is TRUE (default: 4).}
}
\value{
A data frame containing enriched GO terms with associated statistics.
}
\description{
Perform Gene Ontology Enrichment Analysis. This function performs Gene Ontology (GO)
enrichment analysis using the enrichGO function from the clusterProfiler package.
It retrieves enriched GO terms based on a given gene list and specified parameters.
}
\details{
The function uses enrichGO to perform GO enrichment analysis based on the specified parameters.
It optionally groups similar GO terms if groupGO is TRUE and filters significantly enriched terms
based on the specified p-value cutoff if signif is TRUE. The function utilizes org.Hs.eg.db for
gene annotation in human species.
}
\examples{
\dontrun{
# Example usage of enrich.GO function
# Assuming gene_list is a vector of gene symbols or IDs
enrich.GO(gene_list)
go_grouped_signif <- enrich.GO(gene_list = gene_list, pval = 0.05, keyType = "SYMBOL")
go_grouped <- enrich.GO(gene_list = gene_list, pval = 0.05, signif = F)
go_results <- enrich.GO(gene_list = gene_list, pval = 0.05, signif = F, groupGO = F)
go_signif <- enrich.GO(gene_list = gene_list, pval = 0.05, signif = T, groupGO = F)
}

}
