% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrAnalysis.R
\name{enrich.kegg}
\alias{enrich.kegg}
\title{enrich.kegg}
\usage{
enrich.kegg(gene_list, keyType = "SYMBOL", pval = 0.05, signif = TRUE)
}
\arguments{
\item{gene_list}{Character vector of gene symbols or IDs for which KEGG pathway enrichment analysis is performed.}

\item{keyType}{Character specifying the type of gene identifier used in gene_list (default: "SYMBOL").}

\item{pval}{Numeric specifying the p-value cutoff for enriched KEGG pathways (default: 0.05).}

\item{signif}{Logical indicating whether to filter significantly enriched KEGG pathways based on p-value cutoff (default: TRUE).}
}
\value{
A data frame containing enriched KEGG pathways with associated statistics.
}
\description{
Perform KEGG Pathway Enrichment Analysis. This function performs KEGG pathway enrichment
analysis using the enrichKEGG function from the clusterProfiler package. It retrieves enriched KEGG
pathways based on a given gene list and specified parameters.
}
\details{
The function maps gene symbols to entrez IDs using org.Hs.eg.db, performs KEGG pathway enrichment analysis
using enrichKEGG, and optionally filters significantly enriched pathways based on the specified p-value cutoff.
The organism used is "hsa" (Homo sapiens).
}
\examples{
\dontrun{
# Example usage of enrich.kegg function
# Assuming gene_list is a vector of gene symbols or IDs
kegg_res_signif <- enrich.kegg(gene_list = gene_list, signif = TRUE)
}

}
