% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConsensus.R
\name{getConsensus}
\alias{getConsensus}
\title{getConsensus}
\usage{
getConsensus(
  x,
  n.min = NULL,
  group1 = NULL,
  group2 = NULL,
  meth1 = NULL,
  meth2 = NULL,
  meth.comb = NULL,
  exclude = NULL
)
}
\arguments{
\item{x}{An object of class runClassifiers.obj or a binary df.}

\item{n.min}{An integer specifying the minimum number of algorithms in which a gene must be present to be considered as a consensus gene.}

\item{group1}{A character vector specifying the algorithms to be considered in the first group.}

\item{group2}{A character vector specifying the algorithms to be considered in the second group.}

\item{meth1}{A character string specifying the method to be used to compute the consensus genes in group1. Possible values are 'intersect' and 'union'.}

\item{meth2}{A character string specifying the method to be used to compute the consensus genes in group2. Possible values are 'intersect' and 'union'.}

\item{meth.comb}{A character string specifying the method to be used to combine the consensus genes from group1 and group2. Possible values are 'intersect' and 'union'.}

\item{exclude}{A character vector specifying the algorithms to be excluded from the analysis.}
}
\value{
A list containing the consensus genes, the binary data frame,
        and the input parameters.
}
\description{
This function computes the consensus genes from a runClassifiers object or a binary data frame.
After consensus genes are selected and if x is an object of class runClassifiers.obj it also runs the testConsensus
function on the adjusted data stored in the runClassifiers.obj.
}
\details{
The function computes the consensus genes from a runClassifiers object or a binary data frame.
The consensus genes can be choosen to be the genes that are present in at least n.min algorithms,
or the genes that are present in the algorithms specified in group1,
or the genes that are present in the algorithms specified in group1 and group2.
The consensus genes can be computed using the 'intersect' or 'union' methods.
}
\examples{
\dontrun{
# consensus genes are the genes that are present in at least 3 algorithms
cons.1 <- getConsensus(runClassifiers.obj, n.min = 3)
# consensus genes are the genes that are present in both alg1 and alg2
cons.2 <- getConsensus(runClassifiers.obj, group1 = c("alg1", "alg2"), meth1 = "intersect")
# consensus genes are the genes that are present in alg1 and alg2 and alg3 and alg4
cons.3 <- getConsensus(runClassifiers.obj,
  group1 = c("alg1", "alg2"), group2 = c("alg3", "alg4"),
  meth1 = "intersect", meth2 = "intersect", meth.comb = "intersect"
)
}

}
