% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_testConsensus.R
\name{heatmap_plot_heatmaply}
\alias{heatmap_plot_heatmaply}
\title{Heatmap Calculation with Hierarchical Clustering}
\usage{
heatmap_plot_heatmaply(
  heatmap_data,
  colors = (grDevices::colorRampPalette(c("#440154FF", "#3B528BFF", "#21908CFF",
    "#5DC863FF", "#FDE725FF")))(64),
  dendrogram = "both",
  show_dendrogram = c(FALSE, TRUE),
  scale = "row",
  custom_colors = c(`1` = "#2e1457", `0` = "#66a182"),
  margins = c(60, 100, 40, 20),
  grid_color = "white",
  grid_width = 1e-05,
  branches_lwd = 0.4,
  fontsize_row = 12,
  fontsize_col = 5,
  scale_fill_gradient_fun = NULL,
  limits = NULL,
  na.value = "grey50",
  cellnote = NULL,
  cellnote_size = NULL,
  cellnote_textposition = "middle center",
  key.title = "",
  key.xlab = "Value",
  key.ylab = "Frequency"
)
}
\description{
Computes a hierarchical clustering heatmap matrix for a set of consensus genes in a gene expression data frame.
}
