% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pca.consensus}
\alias{pca.consensus}
\title{Perform PCA on Consensus Genes and Generate Plots}
\usage{
pca.consensus(df, cons_genes, class)
}
\arguments{
\item{df}{A data frame containing gene expression data, where rows represent samples and columns represent genes. The column names must match the gene identifiers used in the `cons_genes` vector.}

\item{cons_genes}{A character vector of gene names to be included in the PCA analysis. These are the consensus genes used for the PCA computation.}

\item{class}{A factor or vector representing the class labels for the samples. This is used to color the points in the PCA plot and differentiate the sample groups.}
}
\value{
A combined plot containing two ggplot objects: one showing the PCA scores for PC1 and PC2, and another displaying the top 15 genes with the highest loadings for both principal components.
}
\description{
This function performs Principal Component Analysis (PCA) on a given dataframe of gene expression data, specifically focusing on a subset of consensus genes. It generates two plots: one for PCA scores (PC1 vs. PC2) and another showing the loadings of the top 15 genes for the first two principal components.
}
\details{
The function first filters the gene expression data to include only the consensus genes provided. It then performs PCA on this subset of genes, extracting eigenvalues, scores, and loadings. The PCA scores are plotted to show how samples group based on their principal component values, colored by their class labels. The top 15 genes with the highest loadings on PC1 and PC2 are also plotted as vectors in a second plot. The function utilizes the `ggplot2`, `cli`, and `reshape2` packages to generate the plots.
}
\examples{
\dontrun{
# Example usage:
# Assuming `df` is a data frame of gene expression,
# `cons_genes` is a vector of consensus gene names,
# and `class_labels` is a factor of class labels for each sample.
pca_plot <- pca.consensus(
    df = gene_expression_data,
    cons_genes = consensus_gene_list, class = class_labels
)
print(pca_plot)
}

}
