% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_testConsensus.R
\name{pca_calculation}
\alias{pca_calculation}
\title{Principal Component Analysis (PCA) Calculation}
\usage{
pca_calculation(df, cons_genes, top_loadings = 15)
}
\arguments{
\item{df}{A data frame where rows represent samples and columns represent genes.}

\item{cons_genes}{A character vector containing the names of the consensus genes to be used in the PCA.}

\item{top_n}{An integer specifying the number of top genes to select based on the absolute loading values. Default is 15.}
}
\value{
A list with the following components:
  - `scores`: A data frame with PCA scores for each sample.
  - `loadings`: A data frame with PCA loadings for each gene.
  - `explained_variance`: A data frame containing the variance explained by each principal component.
  - `top_loadings`: A data frame with the top N genes with the highest absolute loadings.
}
\description{
Computes the PCA on a given data frame containing gene expression counts, selecting the top N genes based on absolute loadings.
}
\details{
This function performs PCA using the prcomp function from the stats package. The input data frame is filtered to keep only the specified consensus genes.
The output includes PCA scores, loadings, explained variance, and the top N genes with the highest loadings.
}
\examples{
\dontrun{
pca_results <- pca_calculation(df = expression_data, cons_genes = gene_list, top_n = 20)
}

}
