% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{performances.plot}
\alias{performances.plot}
\title{performances.plot}
\usage{
performances.plot(performances)
}
\arguments{
\item{performances}{A data frame containing the performances of different models.}
}
\value{
A bar plot visualizing the performances of different models.
}
\description{
This function generates a bar plot to visualize the performances of different models.
}
\examples{
\dontrun{
# Example usage of performances.plot function
performances.plot(performances)
}

}
