% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppiNetwork.R
\docType{class}
\name{ppiNetwork.obj-class}
\alias{ppiNetwork.obj-class}
\title{Class ppiNetwork.obj}
\description{
A class to store the results of the ppiNetwork function, including the network graph, initial interactions, new interactions, final genes, and enrichment results.
}
\section{Slots}{

\describe{
\item{\code{graph}}{An object of class `igraph` representing the protein-protein interaction network.}

\item{\code{initialInteractions}}{A `data.frame` containing the initial interactions among the input genes.}

\item{\code{newInteractions}}{A `data.frame` containing the new interactions among the input genes and their neighbors.}

\item{\code{finalGenes}}{A `data.frame` containing the final list of genes with their STRING identifiers and descriptions.}

\item{\code{enrichmentResults}}{A `list` containing the enrichment analysis results for each cluster and the total network.}

\item{\code{inputParams}}{A list of input parameters used to generate the ppiNetwork result.}
}}

