% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppiNetwork.R
\name{ppiNetwork}
\alias{ppiNetwork}
\title{ppiNetwork}
\usage{
ppiNetwork(
  gene_list,
  version = "12",
  species = 9606,
  score_threshold = 400,
  max_interactors = 30,
  frac = 8,
  layout = "stress",
  circular = FALSE,
  palette = "Paired",
  max.overlaps = 20,
  cluster.method,
  bundling = TRUE,
  size_label = 2
)
}
\arguments{
\item{gene_list}{A character vector of gene identifiers.}

\item{version}{A character string specifying the STRING database version. Default is "12".}

\item{species}{An integer specifying the NCBI taxonomy ID of the species. Default is 9606 (Homo sapiens).}

\item{score_threshold}{A numeric value indicating the minimum interaction score for STRING. Default is 400.}

\item{max_interactors}{An integer specifying the maximum number of interactors to include. Default is 30.}

\item{frac}{A numeric value indicating the minimum degree threshold to filter nodes. Default is 8.}

\item{layout}{A character string specifying the layout for network visualization. Default is "stress".}

\item{circular}{A logical value indicating whether to plot the network in a circular layout. Default is FALSE.}

\item{palette}{A character string specifying the color palette for clusters. Default is 'Paired'.}

\item{max.overlaps}{An integer specifying the maximum number of overlaps for node labels. Default is 20.}

\item{cluster.method}{A character string specifying the method for community detection.
Options include "leiden", "edge_betweenness", "infomap", "label_prop", "leading_eigen", "spinglass", "walktrap". Default is "louvain".}

\item{bundling}{Logical. Wether or not bundling edges. Default is TRUE.}

\item{size_label}{Label size. Default is 2.}
}
\value{
An object of class `ppiNetwork` containing the PPI network, initial and
new interactions, final genes, and enrichment results (only significant ones).
}
\description{
Protein-Protein Interaction Network Analysis. This function builds a protein-protein
interaction (PPI) network from a list of genes, finds communities in the Network, and performs enrichment
analysis for each cluster.
}
\examples{
\dontrun{
gene_list <- c("BRCA1", "TP53", "EGFR", "MYC")
result <- ppiNetwork(gene_list, cluster.method = "louvain")
}

}
