% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_preProcess.R
\docType{class}
\name{preProcess.obj-class}
\alias{preProcess.obj-class}
\title{class preProcess.obj}
\description{
Creates a class for data import that will be used to stock pre-processed data.
}
\section{Slots}{

\describe{
\item{\code{raw}}{A data frame containing the raw data.}

\item{\code{processed}}{A list containing the processed data.}

\item{\code{metadata}}{A data frame containing the metadata.}

\item{\code{data.info}}{A list containing the data type and normalization status.}
}}

