% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcess.R
\name{preProcess}
\alias{preProcess}
\title{Pre-process data}
\usage{
preProcess(
  df.count,
  df.clin,
  class = "class",
  case.label = NULL,
  mincpm = 1,
  minfraction = 0.1,
  data.type = "rnaseq",
  is.normalized = FALSE,
  batch = NULL,
  covar.mod = NULL,
  plot = TRUE
)
}
\arguments{
\item{df.count}{A data frame containing the count data. Rows are samples and columns are genes.}

\item{df.clin}{A data frame containing the clinical data. Rows are samples and columns are clinical variables.}

\item{class}{A character string specifying the column name in df.clin that contains the class labels.}

\item{case.label}{A character string specifying the case label in the class column. Default is NULL.}

\item{mincpm}{An integer specifying the minimum count per million (CPM) value for filtering genes. Default is 1.}

\item{minfraction}{A numeric value specifying the minimum fraction of samples a gene must be present in to be retained. Default is 0.1.}

\item{data.type}{A character string specifying the data type. Possible values are 'rnaseq' and 'array'. Default is 'rnaseq'.}

\item{is.normalized}{A logical value specifying if the data is already normalized. Default is FALSE.}

\item{batch}{A character string specifying the column name in df.clin that contains the batch variable. Default is NULL.}

\item{covar.mod}{A character string or a character vector specifying the column name(s) in df.clin that contains the covariate(s) for batch correction. Default is NULL.}

\item{plot}{A logical value specifying if PCA and PVCA plots should be generated. Default is TRUE.}
}
\value{
An object of class preProcess.obj containing the pre-processed data.
}
\description{
This function performs data pre-processing steps including
data normalization, batch correction, and PCA and PVCA plots before and after
batch correction.
}
\details{
The function performs the following steps:
1. Import the data and match samples in both data tables.
2. Normalize the data if the data type is RNA-seq and the data is not normalized.
3. Perform batch correction if the batch variable is provided.
4. Generate PCA and PVCA plots before and after batch correction.
}
\examples{
/dontrun{
preProcess(df.count, df.clin, class = "class",
           case.label = "case", mincpm = 1, minfraction = 0.1,
           data.type = "rnaseq", is.normalized = FALSE,
           batch = "batch", covar.mod = "covar", plot = TRUE)}

}
