% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_runClassifiers.R
\docType{class}
\name{runClassifiers.obj-class}
\alias{runClassifiers.obj-class}
\title{Class for the runClassifiers object}
\description{
Create a class for the runClassifiers object.
}
\section{Slots}{

\describe{
\item{\code{data}}{A list containing the adjusted data and metadata.}

\item{\code{models.info}}{A list containing the finalized workflows for each model.}

\item{\code{model.features}}{A list containing the variable importances for each model.}

\item{\code{performances}}{A list containing the tuning and final metrics for each model.}
}}

