% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_runClassifiers.R
\name{runClassifiers}
\alias{runClassifiers}
\title{runClassifiers}
\usage{
runClassifiers(
  preProcess.obj = NULL,
  ...,
  models = c("bag_mlp", "rand_forest", "svm_poly"),
  selector.recipes = c("boruta", "roc", "boruta"),
  tuning.method = "tune_grid",
  n = 5,
  v = 3,
  metric = "accuracy",
  nsim = 2,
  filter = TRUE,
  seed = 123,
  downsample = FALSE,
  plot = TRUE
)
}
\arguments{
\item{preProcess.obj}{An object of class preProcess.}

\item{...}{Arguments passed to the function. If preProcess.obj is not provided, the function looks
for df.count, df.clin and class in ... as specified in the documentation of preProcess function.}

\item{models}{A character vector specifying the models to be used.
Possible values are 'xgboost', 'bag_tree', 'lightGBM', 'pls', 'logistic',
'C5_rules', 'mars', 'bag_mars', 'mlp', 'bag_mlp', 'decision_tree',
'rand_forest', 'svm_linear', 'svm_poly', 'svm_rbf'.}

\item{selector.recipes}{A character vector specifying the selector recipes to be used.
Possible values are 'boruta', 'roc', 'infgain', 'mrmr', 'corr'. Models and selector.recipes will be paired
in the order they are provided. If the number of recipes doesn't match the number of models, the first recipe will be used for all models.}

\item{tuning.method}{A character string specifying the tuning method to be used.
Possible values are 'tune_grid (default)', tune_race_anova', tune_race_win_loss', 'tune_bayes', 'tune_sim_anneal'.}

\item{n}{An integer specifying the number of iterations for the tuning method.}

\item{v}{An integer specifying the number of folds for the cross-validation during the hyperparameters tuning.}

\item{metric}{A character string specifying the metric to be used for tuning.}

\item{nsim}{An integer specifying the number of simulations for the permutation-based VIP.}

\item{filter}{A logical specifying whether to filter genes not annotated in the GO and KEGG databases.}

\item{seed}{An integer specifying the seed for reproducibility.}

\item{downsample}{A logical specifying whether to downsample the majority class.}
}
\value{
An object of class runClassifiers.obj containing:
data: A list containing the adjusted data and metadata.
models.info: A list containing the finalized workflows for each model.
model.features: A list containing the variable importances for each model.
performances: A list containing the tuning and final metrics for each model.
}
\description{
This function run classifiers specified on an object of class preProcess.obj or
on the data provided in the arguments ...
}
\details{
The function runs specified classifiers paired with provided selector.recipes on pre-processed data.
If filter = TRUE, the function first filters the genes that are not annotated in the GO and KEGG databases.
Then, the function tunes and fits the models using the specified tuning method.
Finally it computes the variable importances for each model using the permutation-based VIP,
when the direct VIP computation fails.
}
\examples{
\dontrun{
rc <- runClassifiers(preProcess.obj,
  models = c("bag_mlp", "rand_forest", "svm_poly"),
  selector.recipes = "boruta", tuning.method = "tune_grid", n = 5,
  v = 3, metric = "accuracy", nsim = 2, seed = 123
)
}

}
