% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_testConsensus.R
\name{show,testConsensus.obj-method}
\alias{show,testConsensus.obj-method}
\title{Display Method for testConsensus.obj}
\usage{
\S4method{show}{testConsensus.obj}(object)
}
\arguments{
\item{object}{An object of class `testConsensus.obj`.}
}
\description{
Custom method for displaying the content of a `testConsensus.obj` object.
Provides a clear summary of the stored data, including the input data, PCA results, and AUROC/FC analysis.
}
\details{
The method dynamically adapts to the content of each slot, ensuring a clear display of large data frames (only the first 5 columns and rows) and a compact view of long vectors.
}
\examples{
\dontrun{
# Assuming `consensus_obj` is an existing testConsensus.obj object:
consensus_obj
}

}
