% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testConsensus.R
\name{testConsensus}
\alias{testConsensus}
\title{Test Consensus}
\usage{
testConsensus(df.count, cons_genes, class)
}
\arguments{
\item{df.count}{A data frame containing the counts data}

\item{cons_genes}{A vector list containing consensus genes to test}

\item{class}{The column name of the class}
}
\value{
The PCA plot, AUROC plot, Heatmap plot, and Validation MLP Model plot
}
\description{
This function validates the consensus genes by performing PCA,
AUROC, Heatmap, and Validation MLP Model on the entire dataset.
}
\details{
This function validates the consensus genes by performing PCA, AUROC, Heatmap,
and Validation MLP Model on the same dataset or a new one. The function first
normalizes the data if it is not already normalized. It then applies batch
correction if a batch variable is provided. The PCA analysis is then performed
on the normalized data. The AUROC analysis is then performed on the consensus
genes. The Heatmap analysis is then performed on the consensus genes. Finally,
the Validation MLP Model is performed on the consensus genes.
}
\examples{
/dontrun{
testConsensus(
   df.count, cons_genes, class)}

}
