% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{upset.plot2}
\alias{upset.plot2}
\title{upset.plot2}
\usage{
upset.plot2(data.obj)
}
\arguments{
\item{data.obj}{An object of class ensBP.}
}
\value{
An UpSet plot visualizing the overlap of features across different models.
}
\description{
This function generates an UpSet plot to visualize the overlap of
features across different models using the ComplexUpSet library.
}
\examples{
\dontrun{
# Example usage of upset.plot2 function
upset.plot2(data.obj)}

}
