% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valConsensus.R
\name{validateConsensus}
\alias{validateConsensus}
\title{Validate Consensus}
\usage{
validateConsensus(
  df.count,
  df.clin,
  getConsensus.obj,
  class = "class",
  case.label = NULL,
  data.type = "rnaseq",
  batch = NULL,
  mincpm = 1,
  minfraction = 0.1,
  covar.mod = NULL,
  is.normalized = FALSE
)
}
\arguments{
\item{df.count}{A data frame containing the counts data}

\item{df.clin}{A data frame containing the clinical data}

\item{getConsensus.obj}{An object containing the consensus genes}

\item{class}{The column name of the class}

\item{case.label}{The label of the case}

\item{data.type}{The type of data. Options are "rnaseq" or "microarray"}

\item{batch}{The column name of the batch}

\item{mincpm}{The minimum count per million to keep a gene}

\item{minfraction}{The minimum fraction of samples to keep a gene}

\item{covar.mod}{The covariate model}

\item{is.normalized}{A boolean indicating if the data is normalized}
}
\value{
The PCA plot, AUROC plot, Heatmap plot, and Validation MLP Model plot
}
\description{
This function validates the consensus genes by performing PCA,
AUROC, Heatmap, and Validation MLP Model on new data.
}
\details{
This function validates the consensus genes by performing PCA, AUROC, Heatmap,
and Validation MLP Model on the same dataset or a new one. The function first
normalizes the data if it is not already normalized. It then applies batch
correction if a batch variable is provided. The PCA analysis is then performed
on the normalized data. The AUROC analysis is then performed on the consensus
genes. The Heatmap analysis is then performed on the consensus genes. Finally,
the Validation MLP Model is performed on the consensus genes.
}
\examples{
/dontrun{
validateConsensus(
   df.count = acc.count, df.clin = acc.clin,
   getConsensus.obj = consensus,
   class = "class", batch = "batch_col",
   )}

}
