% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_runClassifiers.R
\name{wrong.preds.plot}
\alias{wrong.preds.plot}
\title{Plot of Model Prediction Errors}
\usage{
wrong.preds.plot(
  predictions_df,
  color_correct = "#BEBEBE",
  color_wrong = "#BD2B48",
  shape_correct = 16,
  shape_wrong = 16
)
}
\arguments{
\item{predictions_df}{A data frame containing prediction results. It must include the following columns:
- `ID`: Unique identifier for each sample.
- `.pred_class`: The predicted class label.
- `class`: The true class label.
- `model`: The name of the model that made the prediction.}

\item{color_correct}{Color for correct predictions (default: "#6c757d").}

\item{color_wrong}{Color for wrong predictions (default: "#D7263D").}

\item{shape_correct}{Point shape for correct predictions (default: 16, circle).}

\item{shape_wrong}{Point shape for wrong predictions (default: 4, X).}
}
\value{
A ggplot2 object displaying a scatter plot where:
- Each point represents a prediction made by a model.
- Correct predictions are shown in gray, and incorrect predictions in red.
- The x-axis only labels misclassified samples.
}
\description{
Creates a scatter plot to visualize correct and incorrect predictions
across multiple models. Wrong predictions are highlighted, and only misclassified sample IDs
are displayed on the x-axis.
}
\details{
The function first compares the predicted class (`.pred_class`) with the true class (`class`)
to determine whether each prediction is correct or wrong. It then highlights incorrect predictions
with a larger point size and assigns unique IDs on the x-axis only for misclassified samples.
}
\examples{
\dontrun{
pp <- preProcess(df.count = acc.count, df.clin = acc.clin,
                 batch = "batch", covar.mod = "covar")
rc <- runClassifiers(pp, models = c("bag_mlp", "rand_forest", "mlp", "C5_rules"),
                     selector.recipes = c("boruta", "roc", "boruta", "boruta"),
                     filter = TRUE, downsample = TRUE)
wrong.preds.plot(rc@predictions)
}

}
