% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc.count.R
\docType{data}
\name{acc.count}
\alias{acc.count}
\title{acc.count}
\format{
79 rows and 20502 columns.
}
\usage{
acc.count
}
\description{
Count table of ACC data. Samples are on rows and features are on columns.
}
\keyword{datasets}
