% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mlp.model.plot}
\alias{mlp.model.plot}
\title{Create a Plot for MLP Model Variable Importance}
\usage{
mlp.model.plot(df, cons_genes, class)
}
\arguments{
\item{df}{A data frame containing the gene expression data. The dataframe must include a column for the class variable (`class`), which is used for stratified sampling.}

\item{cons_genes}{A character vector containing the names of consensus genes to be used for model training.}

\item{class}{A factor or numeric vector representing the class labels. This is the outcome variable used for classification.}
}
\value{
A `ggplot` object showing the scaled variable importance for the MLP model, with an annotation of model performance metrics.
}
\description{
This function tunes and evaluates a multilayer perceptron (MLP) model using a dataset of consensus genes and creates a plot to visualize the variable importance. The plot shows the scaled importance of each variable in the model, and includes key performance metrics (accuracy, AUROC, and Brier score).
}
\examples{
\dontrun{
# Example usage
plot <- mlp.model.plot(df = gene_data, cons_genes = consensus_genes, class = class_labels)
print(plot)
}

}
