% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_runClassifiers.R
\name{performances.plot}
\alias{performances.plot}
\title{Generate a Bar Plot of Model Performances}
\usage{
performances.plot(data.obj)
}
\arguments{
\item{data.obj}{An object of class `runClassifiers.obj`, which contains the performance metrics
under the slot `@performances`.}
}
\value{
A `ggplot2` object representing the bar plot of model performances.
}
\description{
This function creates a bar plot using `ggplot2` to visualize the performance metrics
of different models during tuning and final fitting stages.
}
\details{
The function extracts model performance metrics from `data.obj@performances`, separating
tuning (`tuning_metrics`) and final fitting (`final_metrics`) performances. Metrics such as
accuracy, F1-score, precision, and recall are visualized using grouped bar plots.

The bars are colored differently for tuning and fitting phases.
The models are displayed on the y-axis, while their performance scores are shown on the x-axis.
The plot uses facets to display different metrics.
}
\examples{
\dontrun{
# Example usage of performances.plot
pp <- preProcess(df.count = acc.count, df.clin = acc.clin,
                batch = "batch", covar.mod = "covar")
rc <- runClassifiers(pp, models = c("bag_mlp", "rand_forest", "mlp", "C5_rules"),
                     selector.recipes = c("boruta", "roc", "boruta", "boruta"),
                     filter = TRUE, downsample = TRUE)
performances.plot(rc)
}

}
