% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTopMetrics}
\alias{plotTopMetrics}
\title{Plot the performance metrics for the top gene combinations.}
\usage{
plotTopMetrics(top_results)
}
\description{
Plot the performance metrics for the top gene combinations.
}
