% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_valConsensus.R
\name{plotTopMetrics}
\alias{plotTopMetrics}
\title{Generate Heatmaps of Top Model Metrics}
\usage{
plotTopMetrics(top_results)
}
\arguments{
\item{top_results}{A data frame containing the results of multiple models with
different gene combinations.}
}
\value{
A `ggplot2` object representing the heatmap of model performances.
}
\description{
This function creates a heatmap using `ggplot2` to visualize the performance
of different clustering and dimensionality reduction models across various metrics.
}
\examples{
\dontrun{
# Example usage of plotTopMetrics
vc <- valConsensus(df.count = count_table, gene.list = gene_list,
                   class = labels, N = 10, metric = "FScore")
p <- plotTopMetrics(vc$topCombinations)
print(p)
}

}
