% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valConsensus.R
\name{selectTopCombinations}
\alias{selectTopCombinations}
\title{Helper function to plot the performance metrics for the top gene combinations}
\usage{
selectTopCombinations(results, N, metric)
}
\description{
Helper function to plot the performance metrics for the top gene combinations
}
