% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testConsensus.R
\name{testConsensus}
\alias{testConsensus}
\title{Test Consensus}
\usage{
testConsensus(df.count, gene.list, class)
}
\arguments{
\item{df.count}{A data frame containing the counts data. Gene names on columns and samples on rows.}

\item{gene.list}{A vector list containing consensus genes to test.}

\item{class}{A vector containing the true class labels.}
}
\value{
The PCA plot, AUROC plot, Heatmap plot, and Validation MLP Model plot
}
\description{
This function tests a list of consensus genes on a df.count performing PCA,
AUROC, Heatmap, and tuning and fitting a MLP Model.
}
\details{
This function tests the consensus genes by performing PCA, AUROC, Heatmap,
and Validation MLP Model on the provided dataset. PCA is performed on the consensus genes,
AUROC is computed for each gene, Heatmap is created using the consensus genes,
and a MLP model is tuned and fitted on the filtered dataset.
}
\examples{
\dontrun{
testConsensus(
    df.count, cons_genes, class
)
}

}
