% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_runClassifiers.R
\name{upset.plot}
\alias{upset.plot}
\title{Generate an Upset Plot for Feature Overlaps Across Models}
\usage{
upset.plot(data.obj)
}
\arguments{
\item{data.obj}{An object of class `runClassifiers.obj`, which contains the selected features for each model
under the slot `@model.features`.}
}
\value{
A `ggplot2` object representing the Upset plot.
}
\description{
This function creates an Upset plot using the `ComplexUpset` package to visualize the overlap
of selected features across multiple models stored in a `runClassifiers.obj` object.
}
\details{
The function extracts features from `data.obj@model.features` and constructs a membership matrix
indicating the presence or absence of each feature across models. The resulting Upset plot
displays the intersection sizes and set sizes, highlighting the most frequently selected features.
}
\examples{
\dontrun{
# Example usage of upset.plot
pp <- preProcess(
    df.count = acc.count, df.clin = acc.clin,
    batch = "batch", covar.mod = "covar"
)
rc <- runClassifiers(pp,
    models = c("bag_mlp", "rand_forest", "mlp", "C5_rules"),
    selector.recipes = c("boruta", "roc", "boruta", "boruta"),
    filter = TRUE, downsample = TRUE
)
upset.plot(rc)
}

}
