% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{upset.plot}
\alias{upset.plot}
\title{upset.plot2}
\usage{
upset.plot(data.obj)
}
\arguments{
\item{data.obj}{An object of class ensBP.}
}
\value{
An UpSet plot visualizing the overlap of features across different models.
}
\description{
This function generates an UpSet plot to visualize the overlap of
features across different models using the ComplexUpSet library.
}
\details{
#' @title upset.plot
#' @description This function generates an UpSet plot to visualize the overlap of
#' features across different models.
#'
#' @param data.obj An object of class ensBP.
#'
#' @return An UpSet plot visualizing the overlap of features across different models.
#'
#' @import ggplot2
#' @importFrom UpSetR fromList upset
#'
#' @examples
#' \dontrun{
#' # Example usage of upset.plot function
#' upset.plot(data.obj)}
#'
#' @export
upset.plot <- function(data.obj) { # TODO fix upset plot
    model_features_list <- data.obj@model.features
    features <- lapply(model_features_list, function(x) unique(x$Variable))
    names(features) <- names(model_features_list)
    upset_list <- UpSetR::fromList(features)

    color <- "steelblue"

    p <- UpSetR::upset(upset_list,
        order.by = "freq",
        keep.order = TRUE,
        nsets = length(upset_list),
        nintersects = NA,
        sets.x.label = "Set size",
        mainbar.y.label = "Intersection size",
        point.size = 4,
        line.size = 1.5,
        text.scale = c(2.8, 2.4, 2.3, 2.4, 2.4, 2.4),
        # text.scale = c(5.8, 5.4, 5.3, 5.4, 5.4, 5.4),
        mb.ratio = c(0.6, 0.4),
        matrix.color = color,
        sets.bar.color = color,
        main.bar.color = color,
        shade.color = color,
        shade.alpha = 0.09,
    )

    return(p)
}
}
\examples{
\dontrun{
# Example usage of upset.plot2 function
upset.plot(data.obj)}

}
